+++ weyder.js
@@ -2,11 +2,11 @@ const https = require('https');
 
 const {MapBoxObject,DarkSkyObject,buildURL} = require('./lib/builder.js')
 
-
 const geoCode = (address,callback) => {
     MapBoxObject.urlComponent.place = address
     if(callback == null || callback == undefined && typeof(callback)!=='function'){
         return new Promise((resolve,reject)=>{
+            if(MapBoxObject.queryStringObj.accessToken == '') throw new Error('Please provide mapbox access token')
             https.get(buildURL('mapbox'), (resp) => {
                 let data = '';
                 resp.on('data', (chunk) => {
@@ -33,6 +33,7 @@ const geoCode = (address,callback) => {
             });
         })
     }else{
+        if(MapBoxObject.queryStringObj.accessToken == '') throw new Error('Please provide mapbox access token')
         https.get(buildURL('mapbox'), (resp) => {
             let data = '';
             resp.on('data', (chunk) => {
@@ -65,7 +66,7 @@ const foreCast = ({latitude,longitude,place_name:place},callback)=>{
     DarkSkyObject.urlComponent.long = longitude
     if(callback == null || callback == undefined && typeof(callback)!=='function'){
         return new Promise((resolve,reject)=>{
-            console.log(buildURL('darksky'));
+            if(DarkSkyObject.urlComponent.accessToken == '') throw new Error('Please provide darksky access token')
             https.get(buildURL('darksky'), (resp) => {
                 let data = '';
                 resp.on('data', (chunk) => {
@@ -77,12 +78,12 @@ const foreCast = ({latitude,longitude,place_name:place},callback)=>{
                         reject(`Server responded with ${allData.code} ${allData.error}`)
                         
                     }else{
-                        let {summary,temperature,precipProbability:rainChance} = allData.currently
+                        console.log(allData);
+                        let {currently={},hourly={},daily={}} = allData
                         let forecastInfo = {
-                            summary,
-                            place,
-                            temperature:temperature+'°C',
-                            rainChance: (rainChance*100).toFixed(2)+'%'
+                            currently,
+                            hourly,
+                            daily
                         }
                         resolve(forecastInfo)
                     }
@@ -90,6 +91,7 @@ const foreCast = ({latitude,longitude,place_name:place},callback)=>{
             }).on('error', (e) => {reject(e)});
         })
     }else{
+        if(DarkSkyObject.urlComponent.accessToken == '') throw new Error('Please provide darksky access token')
         https.get(buildURL('darksky'), (resp) => {
             let data = '';
             resp.on('data', (chunk) => {
@@ -100,12 +102,11 @@ const foreCast = ({latitude,longitude,place_name:place},callback)=>{
                 if(allData.error){
                     callback(`Server responded with ${allData.code} ${allData.error}`,undefined)
                 }else{
-                    let {summary,temperature,precipProbability:rainChance} = allData.currently
+                    let {currently={},hourly={},daily={}} = allData
                     let forecastInfo = {
-                        summary,
-                        place,
-                        temperature:temperature+'°C',
-                        rainChance: (rainChance*100).toFixed(2)+'%'
+                        currently,
+                        hourly,
+                        daily
                     }
                     callback(undefined,forecastInfo)
                 }
@@ -118,9 +119,18 @@ const setAccessToken = function(obj,token){
     if(obj == 'darksky') DarkSkyObject.urlComponent.accessToken = token 
     else if(obj == 'mapbox') MapBoxObject.queryStringObj.access_token = token
 }
+const setWeatherDataSpan = function(exObj){
+    let allowableEx = ['currently','hourly','daily']
+    let excluded = DarkSkyObject.queryStringObj.exclude
+    Object.entries(exObj).forEach((item,index)=>{
+        if(item[1]==true) excluded.splice(DarkSkyObject.queryStringObj.exclude.indexOf(item[0]),1)
+    })
+    DarkSkyObject.queryStringObj.exclude = excluded
+}
 
 module.exports = {
     geoCode,
     foreCast,
-    setAccessToken
+    setAccessToken,
+    setWeatherDataSpan
 };